Remove methods, that have been moved to PowerSystemUtils
