+++ src/main/java/edu/ie3/datamodel/utils/GridAndGeoUtils.java
@@ -5,20 +5,11 @@
 */
 package edu.ie3.datamodel.utils;
 
-import static edu.ie3.util.quantities.PowerSystemUnits.*;
-
 import edu.ie3.datamodel.models.input.NodeInput;
 import edu.ie3.util.geo.GeoUtils;
-import java.util.*;
-import java.util.stream.Collectors;
 import javax.measure.quantity.Length;
-import org.apache.commons.lang3.ArrayUtils;
-import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.LineString;
-import org.locationtech.jts.geom.Point;
-import org.locationtech.jts.geom.impl.CoordinateArraySequence;
 import tech.units.indriya.ComparableQuantity;
-import tech.units.indriya.quantity.Quantities;
 
 /** This class offers some useful methods for handling geographical problems related to grids */
 public class GridAndGeoUtils extends GeoUtils {
@@ -54,100 +45,6 @@ public class GridAndGeoUtils extends GeoUtils {
   }
 
   /**
-   * Build an instance of {@link LineString} between two points that is safe to be compared even if
-   * the provided two points consist of exactly the same coordinates. This is done by increasing the
-   * coordinate of the provided Point {@code p1} by a small amount to make it different from Point
-   * {@code p2}. For details on the bug inside {@link LineString} that is addressed here, see
-   * https://github.com/locationtech/jts/issues/531
-   *
-   * @param p1 start point of the linestring
-   * @param p2 end point of the linestring
-   * @return a {@link LineString} between the provided points
-   */
-  public static LineString buildSafeLineStringBetweenPoints(final Point p1, final Point p2) {
-    final Point safePoint1 = p1.equals(p2) ? buildSafePoint(p1) : p1;
-    return DEFAULT_GEOMETRY_FACTORY.createLineString(
-        ArrayUtils.addAll(safePoint1.getCoordinates(), p2.getCoordinates()));
-  }
-
-  /**
-   * Build an instance of {@link LineString} between two coordinates that is safe to be compared
-   * even if the provided two coordinates are exactly the same coordinates. This is done by
-   * increasing the coordinate of the provided Point {@code c1} by a small amount to make it
-   * different from Point {@code c2}. For details on the bug inside {@link LineString} that is
-   * addressed here, see https://github.com/locationtech/jts/issues/531
-   *
-   * @param c1 start coordinate of the linestring
-   * @param c2 end coordinate of the linestring
-   * @return A safely build line string
-   */
-  public static LineString buildSafeLineStringBetweenCoords(
-      final Coordinate c1, final Coordinate c2) {
-    final Coordinate safeCoord1 = c1.equals(c2) ? buildSafeCoord(c1) : c1;
-    return DEFAULT_GEOMETRY_FACTORY.createLineString(
-        ArrayUtils.addAll(new Coordinate[] {safeCoord1}, c2));
-  }
-
-  /**
-   * Convert a given {@link LineString} with at least two points into a 'safe to be compared' {@link
-   * LineString} This is done by removing duplicates in the points in the provided linestring as
-   * well as a small change of the start coordinate if the linestring only consists of two
-   * coordinates. For details on the bug inside {@link LineString} that is addressed here, see
-   * https://github.com/locationtech/jts/issues/531
-   *
-   * @param lineString the linestring that should be checked and maybe converted to a 'safe to be
-   *     compared' linestring
-   * @return a 'safe to be compared' linestring
-   */
-  public static LineString buildSafeLineString(LineString lineString) {
-    if (lineString.getCoordinates().length == 2) {
-      return buildSafeLineStringBetweenPoints(lineString.getStartPoint(), lineString.getEndPoint());
-    } else {
-      // rebuild line with unique points
-      /* Please note, that using a simple HashSet here to obtain uniqueness is harmful, as it not necessarily maintains
-       * the order of coordinates (but most likely will). Additionally, the behaviour of a HashSet might change with the
-       * JVM (version) you use to execute the code. */
-      Coordinate[] uniqueCoords =
-          Arrays.stream(lineString.getCoordinates()).distinct().toArray(Coordinate[]::new);
-      return uniqueCoords.length == 1
-          ? buildSafeLineStringBetweenPoints(lineString.getStartPoint(), lineString.getEndPoint())
-          : DEFAULT_GEOMETRY_FACTORY.createLineString(uniqueCoords);
-    }
-  }
-
-  /**
-   * Adapted {@link Coordinate#x}, {@link Coordinate#y} and {@link Coordinate#z} of the provided
-   * {@link Coordinate} by 1e-13 and return a new, adapted instance of {@link Coordinate}
-   *
-   * @param coord the coordinate that should be adapted
-   * @return the adapted coordinate with slightly changed x,y,z values
-   */
-  private static Coordinate buildSafeCoord(Coordinate coord) {
-
-    double modVal = 1e-13;
-    double p1X = coord.getX() + modVal;
-    double p1Y = coord.getY() + modVal;
-    double p1Z = coord.getZ() + modVal;
-
-    return new Coordinate(p1X, p1Y, p1Z);
-  }
-
-  /**
-   * Adapt the provided point as described in {@link #buildSafeCoord(Coordinate)} and return a new,
-   * adapted instance of {@link Point}
-   *
-   * @param p1 the point that should be adapted
-   * @return the adapted point with a slightly changed coordinate
-   */
-  private static Point buildSafePoint(Point p1) {
-
-    Coordinate[] safeCoord = new Coordinate[] {buildSafeCoord(p1.getCoordinate())};
-    CoordinateArraySequence safeCoordSeq = new CoordinateArraySequence(safeCoord);
-
-    return new Point(safeCoordSeq, p1.getFactory());
-  }
-
-  /**
    * Calculates the distance between two {@link NodeInput} entities using {@link
    * #calcHaversine(double, double, double, double)}
    *
@@ -161,41 +58,5 @@ public class GridAndGeoUtils extends GeoUtils {
         nodeA.getGeoPosition().getX(),
         nodeB.getGeoPosition().getY(),
         nodeB.getGeoPosition().getX());
-  }
-
-  /**
-   * Calculates and sorts the distances between a base coordinate and other given coordinates using
-   * {@link #calcHaversine(double, double, double, double)}
-   *
-   * @param baseCoordinate the base point
-   * @param coordinates the points to calculate the distance from the base point for
-   * @return a sorted set of distances between the base and other coordinates
-   */
-  public static SortedSet<CoordinateDistance> getCoordinateDistances(
-      Point baseCoordinate, Collection<Point> coordinates) {
-    return coordinates.stream()
-        .map(coordinate -> new CoordinateDistance(baseCoordinate, coordinate))
-        .collect(Collectors.toCollection(TreeSet::new));
-  }
-
-  /**
-   * Calculates the total length of a LineString through building the sum of the distances between
-   * all points of LineString using {@link #calcHaversine(double, double, double, double)}
-   *
-   * @param lineString the line string which length shall be calculated
-   * @return The total length of the line string
-   */
-  public static ComparableQuantity<Length> totalLengthOfLineString(LineString lineString) {
-    ComparableQuantity<Length> y = Quantities.getQuantity(0, KILOMETRE);
-    for (int i = 0; i < lineString.getNumPoints() - 1; i++) {
-      y =
-          y.add(
-              calcHaversine(
-                  lineString.getCoordinateN(i).getX(),
-                  lineString.getCoordinateN(i).getY(),
-                  lineString.getCoordinateN(i + 1).getX(),
-                  lineString.getCoordinateN(i + 1).getY()));
-    }
-    return y;
   }
 }
